move BEFORE_APP_SERIALIZED hook into the lib
